Refactor show_vm_templates to DCLCalculatorPluginModel from VMPricing
